requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
transformers
datasets
tqdm
streamlit
tweepy
botometer
networkx
python-dotenv
xgboost
plotly
shap
torch
jupyterlab
wordcloud
textblob
umap-learn
nltk
sentence-transformers